cmd: remove verbose block data
This is a lot harder to visualize with the new torrent-piece backend,
since we don’t keep an explicit array of which blocks are reserved.
Just remove it for now.

If we need this visualization back to debug something, it should be
possible to recreate it from torrent-piece with a bit of work.